// frontend/src/types/types.ts
export type IAddress = {
  address: string;
  lat: number;
  lng: number;
};
export type VehicleTypeKeys =
  | "Standard Car"
  | "Large Car"
  | "Extra-Large Car"
  | "Electric Car"
  | "Large Electric Car"
  | "Taxi"
  | "Small Truck"
  | "Medium Truck"
  | "Large Truck"
  | "Small Bus"
  | "Large Bus"
  | "Motorcycle"
  | "Large Motorcycle"
  | "Small RV"
  | "Medium RV"
  | "Large RV";